impr: atuo-complete variables from git repo
