+++ libs/domains/organizations/feature/src/index.ts
@@ -80,3 +80,4 @@ export * from './lib/hooks/use-container-versions/use-container-versions'
 export * from './lib/hooks/use-helm-repository-associated-services/use-helm-repository-associated-services'
 export * from './lib/hooks/use-container-registry-associated-services/use-container-registry-associated-services'
 export * from './lib/hooks/use-organization-credentials/use-organization-credentials'
+export * from './lib/hooks/use-parse-terraform-variables-from-git-repo/use-parse-terraform-variables-from-git-repo'
