+++ libs/domains/service-terraform/feature/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/hooks/use-create-terraform-service/use-create-terraform-service'
 export * from './source-setting/source-setting'
-export * from './values-override-arguments-setting/values-override-arguments-setting'
